Feat - submit preprint  (#169)
* refactor(preprints-state): Renamed Preprints -> PreprintProviders

* refactor(preprints-services): Renamed services to have 'preprints' prefix. Decomposed

* fix(json-api-post): Fixed errors during to merge conflict

* refactor(preprints-state): Renamed Preprints -> PreprintProviders

* refactor(preprints-models): Decomposed, used shared models. Moved some to shared

* refactor(router): Moved preprint routes to root folder of feature

* refactor(preprints-components): Rename and reorganize component files for consistency

* fix(preprints-models): Removed unused model

* refactor(license): Made license more reusable by removing card wrapper

* refactor(metadata-license): Removed redundant wrapper

* fix(comments): Fixed comments

* feat(author-assertions): Partly implemented step with missing available case handling

* refactor(validators): Moved linkValidator from helper function to shared util

* feat(array-input): Implemented reusable component for array input

* feat(array-input): Used array-input for data links and prereg links

* refactor(author-assertions): Refactored step. Fixed minor issues

* fix(translations): Update URL validation message for clarity

* fix(comments): Fixed comments

* refactor(subjects): Refactored shared subjects component and service

* feat(metadata-subjects): Implemented Subjects section on metadata step

* fix(discover-state): Removed selectSnapshot that selects data within the same store

* feat(regions): Introduced shared state for fetching regions

* feat(supplements): Implemented template for 'Choose existing' option

* refactor(add-project-form): Refactored form component to be more reusable

* refactor(node-models): Reorganized shared node models

* feat(supplements-step): Implemented connecting project step

* fix(conflict): Fixed things after merge conflict

* fix(conflict): Fixed things after merge conflict

* fix(registries-subjects): Use constant for OSF provider ID in fetchSubjects

* fix(submit-preprint): Fixes after merge conflict

* fix(author-assertions): Fixed link info dropdown options

* fix(file-step): Using preprint word from provider on UI

* feat(review-step): Implemented UI for review step

* fix(supplement-step): Fixed next button disable status

* fix(submit-stepper): Made steps variable computed based on provider

* refactor(subjects): Refactored subjects validation logic

* feat(review-step): Integrated API to submit preprint

* feat(submit-preprint): Handled loosing unsaved data when moving back and away

* fix(conflict): Fixed after merge conflict

* fix(submit-preprint): Minor fixes as a result of manual testing

* feat(submit-preprint): Extracted static string to en.json

* fix(comments): Moved payload forming to mapper

+++ src/app/features/preprints/store/submit-preprint/submit-preprint.state.ts
@@ -26,6 +26,7 @@ import {
   CreatePreprint,
   DisconnectProject,
   FetchLicenses,
+  FetchPreprintById,
   FetchPreprintProject,
   GetAvailableProjects,
   GetPreprintFiles,
@@ -37,6 +38,7 @@ import {
   SaveLicense,
   SetSelectedPreprintFileSource,
   SetSelectedPreprintProviderId,
+  SubmitPreprint,
   SubmitPreprintStateModel,
   UpdatePreprint,
   UploadFile,
@@ -81,6 +83,7 @@ const DefaultState: SubmitPreprintStateModel = {
     isLoading: false,
     error: null,
   },
+  hasBeenSubmitted: false,
 };
 
 @State<SubmitPreprintStateModel>({
@@ -121,6 +124,18 @@ export class SubmitPreprintState {
     return this.preprintsService.updatePreprint(action.id, action.payload).pipe(
       tap((preprint) => {
         ctx.setState(patch({ createdPreprint: patch({ isSubmitting: false, data: preprint }) }));
+      }),
+      catchError((error) => this.handleError(ctx, 'createdPreprint', error))
+    );
+  }
+
+  @Action(FetchPreprintById)
+  getPreprintById(ctx: StateContext<SubmitPreprintStateModel>, action: FetchPreprintById) {
+    ctx.setState(patch({ createdPreprint: patch({ isLoading: true }) }));
+
+    return this.preprintsService.getById(action.id).pipe(
+      tap((preprint) => {
+        ctx.setState(patch({ createdPreprint: patch({ isLoading: false, data: preprint }) }));
       }),
       catchError((error) => this.handleError(ctx, 'createdPreprint', error))
     );
@@ -285,7 +300,7 @@ export class SubmitPreprintState {
     const state = ctx.getState();
     const createdPreprintId = state.createdPreprint.data?.id;
     ctx.setState({ ...DefaultState });
-    if (createdPreprintId) {
+    if (createdPreprintId && !state.hasBeenSubmitted) {
       return this.preprintsService.deletePreprint(createdPreprintId);
     }
 
@@ -467,6 +482,18 @@ export class SubmitPreprintState {
         }),
         catchError((error) => this.handleError(ctx, 'preprintProject', error))
       );
+  }
+
+  @Action(SubmitPreprint)
+  submitPreprint(ctx: StateContext<SubmitPreprintStateModel>) {
+    const createdPreprintId = ctx.getState().createdPreprint.data!.id;
+    ctx.setState(patch({ createdPreprint: patch({ isSubmitting: true }) }));
+    return this.preprintsService.submitPreprint(createdPreprintId).pipe(
+      tap(() => {
+        ctx.setState(patch({ createdPreprint: patch({ isSubmitting: false }), hasBeenSubmitted: true }));
+      }),
+      catchError((error) => this.handleError(ctx, 'createdPreprint', error))
+    );
   }
 
   private handleError(
